+++ pkgs/jni/lib/src/nio/jbyte_buffer.dart
@@ -40,7 +40,7 @@ final class $JByteBuffer$Type$ extends JType<JByteBuffer> {
 /// such as [Uint8List.setRange].
 ///
 /// Example:
-/// <!-- file://./../../../example/api/jbyte_buffer_snippet.dart#allocate_direct -->
+/// <!-- file://./../../../tool/snippets/jbyte_buffer_snippet.dart#allocate_direct -->
 /// ```dart
 /// final directBuffer = JByteBuffer.allocateDirect(3);
 /// directBuffer.asUint8List().setAll(0, [1, 2, 3]);
@@ -53,7 +53,7 @@ final class $JByteBuffer$Type$ extends JType<JByteBuffer> {
 /// correctly garbage collects the buffer and frees its underlying memory.
 ///
 /// Example:
-/// <!-- file://./../../../example/api/jbyte_buffer_snippet.dart#release_buffer -->
+/// <!-- file://./../../../tool/snippets/jbyte_buffer_snippet.dart#release_buffer -->
 /// ```dart
 /// final directBuffer = JByteBuffer.allocateDirect(3);
 /// final data = directBuffer.asUint8List();
@@ -65,7 +65,7 @@ final class $JByteBuffer$Type$ extends JType<JByteBuffer> {
 /// by setting the `releaseOriginal` parameter to `true`.
 ///
 /// Example:
-/// <!-- file://./../../../example/api/jbyte_buffer_snippet.dart#release_original -->
+/// <!-- file://./../../../tool/snippets/jbyte_buffer_snippet.dart#release_original -->
 /// ```dart
 /// final directBuffer = JByteBuffer.allocateDirect(3);
 /// // [releaseOriginal] is `false` by default.