+++ pkgs/jnigen/lib/src/bindings/dart_generator.dart
@@ -25,7 +25,7 @@ const String version = '1.0.1';
 @visibleForTesting
 const String jniMajorVersion = '1';
 @visibleForTesting
-const String jniMinorVersion = '0';
+const String jniMinorVersion = '1';
 
 // Import prefixes.
 const _jni = r'jni$_';